refactor(deep_causality): promoted CSM modules from files into directories.
